perf(stream): stage 3 = two style slices per worker (drop the run-merge experiment)
The nearly-sorted k-way index merge WON natively (1.30s vs ~2s argsort on
the real 19.1M-id sequence) but REGRESSED in wasm — branchy tuple-compare
merging loses badly to pdqsort there (styles 9.7 -> 10.4-10.7s, first
visible 9.9 -> 11.6-11.9s across 3 runs). Reverted to main's argsort.

What stays is the measured win: 2 style slices per worker, round-robin.
The styles tail is max(worker), and one worker landing on a slow core held
the whole tail hostage; smaller slices halve that damage and interleave
better with the concurrent pre-pass tail + parser.

CatiaA1 (883MB, cold, 3 runs): meta 2.3-2.7s, styles 5.5-5.8s, first
visible 6.8/7.1/7.1s (stage 2: 9.9s; serial: 14.3s). Small-model parity
exact (1960 styled / 62 voids / 83 draws / 399ms).
